Make roomserver use the factored out partition offsets table
